Move all shift methods into shiftable mixin
